Cleanup

+++ fastreach-core/src/lib.rs
@@ -2,4 +2,4 @@ pub mod cascade;
 pub mod graph;
 pub mod vincenty;
 
-pub const MOVE_SPEED: f64 = 1000.0 / 12.0; // in m/min
+pub const MOVE_SPEED: f32 = 1000.0 / 12.0; // in m/min
